requirements.txt
torch==2.2
torchvision==0.17
scikit-learn==1.4.2
numpy==1.20.1
pillow==10.3.0

